requirements.txt
scikit-learn
pandas
pydantic
fastapi
bentoml<1.4
requests
PyJWT==2.6.0
pytest
starlette
